+++ src/ESPAsyncTCP.h
@@ -73,7 +73,7 @@ class AsyncClient {
     void _error(int8_t err);
     int8_t _poll(tcp_pcb* pcb);
     int8_t _sent(tcp_pcb* pcb, uint16_t len);
-    int8_t _recv(tcp_pcb* pcb, pbuf* pb, int8_t err);
+    int32_t _recv(tcp_pcb* pcb, pbuf* pb, int8_t err);
     static int8_t _s_poll(void *arg, struct tcp_pcb *tpcb);
     static int32_t _s_recv(void *arg, struct tcp_pcb *tpcb, struct pbuf *pb, int8_t err);
     static void _s_error(void *arg, int8_t err);
